+++ src/main/java/roomescape/controller/api/MemberAuthController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import static roomescape.controller.TokenExtractor.extractTokenFromCookie;
