+++ src/main/java/roomescape/controller/api/MemberReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import jakarta.validation.Valid;
 import java.net.URI;
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.controller.Auth;
 import roomescape.controller.request.LoginMemberInformation;
 import roomescape.controller.request.MemberReservationWebRequest;
 import roomescape.controller.response.MemberReservationWebResponse;
